Color texture in separate module
